+++ kafka-streams/kafka-streams/src/test/scala/com/twitter/finatra/kafkastreams/integration/finatratransformer/WordLengthServer.scala
@@ -18,7 +18,7 @@ class WordLengthServer extends KafkaStreamsTwitterServer {
   override protected def configureKafkaStreams(streamsBuilder: StreamsBuilder): Unit = {
 
     kafkaStreamsBuilder.addStateStore(
-      FinatraTransformer.timerStore(timerStoreName, Serdes.String()))
+      FinatraTransformer.timerStore(timerStoreName, Serdes.String(), streamsStatsReceiver))
 
     val transformerSupplier = () =>
       new WordLengthFinatraTransformer(statsReceiver, timerStoreName)